requirements.txt
flask
yt-dlp
PyPDF2
pdfminer.six
gunicorn
